Backdrop (#80)
* hardcode in the backdrop

* port over default, current, front panel, back panel

* use didUpdateWidget to forward animate

* dropdown from/to values are taken from widgets.units instead of _from/_toValue

* use Dart 2 for backdrop; correct the dropdown _from/_toValue

* alphabetize imports

* update error conditional

* split Category into data object and widget, add docs to Backdrop, cleanup

* pass in Widgets for titles, use _ for private classes

* rename ConverterRoute to UnitConverter

* code cleanup, better loopings and mapping

* add error for non-Map json response from API

* remove extra SingleChildScrollView

+++ unit_converter/unit_converter/lib/unit.dart
@@ -9,6 +9,7 @@ class Unit {
   final String name;
   final double conversion;
 
+  /// A [Unit] stores its name and conversion factor (compared to a base unit).
   const Unit({
     @required this.name,
     @required this.conversion,
@@ -18,7 +19,7 @@ class Unit {
 
   Unit.fromJson(Map jsonMap) :
     name = jsonMap['name'],
-    conversion = jsonMap['conversion'],
+    conversion = jsonMap['conversion'].toDouble(),
     assert(name != null),
     assert(conversion != null);
 }
